[hotfix] Extract utils of CheckpointCoordinatorTest into a separate utils class and correct codestyle
